Http_enhancements
