+++ src/main/java/io/cdap/plugin/http/common/BaseHttpConfig.java
@@ -44,6 +44,9 @@ public abstract class BaseHttpConfig extends ReferencePluginConfig {
     public static final String PROPERTY_CLIENT_SECRET = "clientSecret";
     public static final String PROPERTY_SCOPES = "scopes";
     public static final String PROPERTY_REFRESH_TOKEN = "refreshToken";
+    public static final String PROPERTY_PROXY_URL = "proxyUrl";
+    public static final String PROPERTY_PROXY_USERNAME = "proxyUsername";
+    public static final String PROPERTY_PROXY_PASSWORD = "proxyPassword";
 
     public static final String PROPERTY_AUTH_TYPE_LABEL = "Auth type";
 
@@ -86,6 +89,24 @@ public abstract class BaseHttpConfig extends ReferencePluginConfig {
     @Description("Endpoint for the resource server, which exchanges the authorization code for an access token.")
     @Macro
     protected String tokenUrl;
+
+    @Nullable
+    @Name(PROPERTY_PROXY_URL)
+    @Description("Proxy URL. Must contain a protocol, address and port.")
+    @Macro
+    protected String proxyUrl;
+
+    @Nullable
+    @Name(PROPERTY_PROXY_USERNAME)
+    @Description("Proxy username.")
+    @Macro
+    protected String proxyUsername;
+
+    @Nullable
+    @Name(PROPERTY_PROXY_PASSWORD)
+    @Description("Proxy password.")
+    @Macro
+    protected String proxyPassword;
 
     @Nullable
     @Name(PROPERTY_CLIENT_ID)
@@ -255,6 +276,21 @@ public abstract class BaseHttpConfig extends ReferencePluginConfig {
         String serviceAccountType = getServiceAccountType();
         return Strings.isNullOrEmpty(serviceAccountType) ? null :
                 serviceAccountType.equals(PROPERTY_SERVICE_ACCOUNT_JSON);
+    }
+
+    @Nullable
+    public String getProxyUrl() {
+        return proxyUrl;
+    }
+
+    @Nullable
+    public String getProxyUsername() {
+        return proxyUsername;
+    }
+
+    @Nullable
+    public String getProxyPassword() {
+        return proxyPassword;
     }
 
     @Nullable
